Added a functionality to check an infinite norm on a function by interval arithmetic. The corresponding command is:
checkinfnorm <function> in <range> bounded by <constant expression> [, diam = <diameter>];

Highly improved the quality of Taylor interval evaluations and thus
improved the quality of the certified infinite norm by allowing the
Taylor evaluation to recurse on the linear term up to a certain
level. The default level is 0, so by default, nothing changes. The
command for setting the recursion level is

taylorrecursions = <positive or zero integer>;


git-svn-id: svn://scm.gforge.inria.fr/svnroot/sollya/trunk@96 af37bc7e-59aa-4313-bbaf-7d19c0032f32