(1)
Corrected a bug in reproducibility of the displaying of values:

For precisions prec set between 32 and 64, integer values between 2^31
and 2^63-1 were displayed differently on machines for which
sizeof(int) = 8 than for those for which sizeof(int) = 16.

(2)

Degraded 4 warnings on D, DD, TD, DE 

"Warning: rounding a value computed on less than 53 bits to double
precision."

to an information.

(3)

Changed the behaviour of the tool on automatic printing (on the
prompt) if no faithful rounding can be determined for a constant
expression.

For constant expressions f(c) for which it can be proved that it
exists two reals (not including infinity) a and b such that 

a <= f(c) <= b 

and a * b < 0, 0 is displayed instead of some FP-determined value if
no faithful rounding for f(c) can be determined.

The warning texts have significantly been changed, hence an update of
the check files. 

The change might involve issue with backward compatibility.

ATTENTION:

The help text and documentation files have not been updated yet. 

THIS MUST BE DONE AS SOON AS POSSIBLE.


(4)

Corrected a bug concerning rationalmode. The value of that global
constant did not get reset to its default value on restart.




git-svn-id: svn://scm.gforge.inria.fr/svnroot/sollya/trunk@809 af37bc7e-59aa-4313-bbaf-7d19c0032f32