* Patched Newton's algorithm: except if there is a bug in the patch, the algorithm should be more stable now. In particular, there was a bug when searching for the zero of f in [a,b] where a<0<b and f(0) = NaN. This bug is now solved. This bug had been introduced in revision 783 and was observed for instance when calling remez((2^x-1)/x,9,[-1/16,1/16]) that had become very slow due to the bug. Christoph also observed the bug when trying to improve dirtyinform.
* I added a new possible returned value to evaluateFaithfulWithCutOffFast. In fact, I used to believe that 0 was returned if
and only if the algorithm was unable to find a precision guaranteeing the faithful rounding. It appears that 0 was also returned
when the algorithm does not manage to evaluate the expression into a real number (e.g. it evaluates to Inf or NaN).
Now, evaluateFaithfulWithcutOffFast returns 3 in this case. This allows one to see the difference between a problem of accuracy
(i.e. f(x) is too small) and a problem of evaluation (i.e. f(x) is not defined or has a false singularity that cannot be detected).
I modified every functions that use the returned value of evaluateFaithfulWithCutOffFast, in order to keep the compatibility.

* I added a test case for remez((2^x-1)/x,9,[-1/16,1/16]) in check.sollya
I also numbered the test cases because it was really a mess to find out where a problem came from when somthing changed
in check.sollya.output.
I also discovered some bugs in check.sollya (syntax errors) that I corrected.
Finally, the test case number 171 invoked 'quit' which prevented to run the following tests. I guess it was a mistake, so
I commented it. If it was intended, please uncomment it, but put this test case as the last one.



git-svn-id: svn://scm.gforge.inria.fr/svnroot/sollya/trunk@866 af37bc7e-59aa-4313-bbaf-7d19c0032f32