Patched a problem with matching with default on the concat operator ("@") and on bounds of ranges ("match [1;2] with [default;2] : (1);")
Corrected a possible bug in bashevaluate if ever the pipe from the
child process that gets forked off to run the bash command gets closed
once before the child process ends, then filled up again and is
eventually closed after child process ends.

Recompiled the PDF documentation from the sources as it was not
up-to-date.

Added additional test lines in the check files and in the memoryleak 
check files to test the behavior on default described above.



git-svn-id: svn://scm.gforge.inria.fr/svnroot/sollya/trunk@1388 af37bc7e-59aa-4313-bbaf-7d19c0032f32