* Change in remez and fpminimax commands syntax: the second argument can now be any list of functions, and allows the user to compute approximation in any basis.
* IMPORTANT NOTE: the signature of the internal functions remez and remezAux
changed. This might affect users of the library. The signature of the 
internal functions FPminimax and FPminimaxMain did not change; however
the meaning of the chain *monomials argument changed: it used to be a 
list of integers and is now a list of pure trees.

* remez now returns a non-hornerized polynomial. However, it is not hard to
maintain compatibility with the old behavior, if needed.

* memoryleaks checks have been added, and the new functionnalities are
documented. However, no test case has been added yet.



git-svn-id: svn://scm.gforge.inria.fr/svnroot/sollya/trunk@1438 af37bc7e-59aa-4313-bbaf-7d19c0032f32