* The patch at revision 2272 did not only fix a bug in the warning but also a double-rounding issue that was present from a very long time ago.
Irony of history: a test was present in the checks precisely for this purpose, but the reference value had not been checked carefully enough and was wrong.

So, to sum up: a bug has been fixed @2272 while we were not even aware of the presence of the bug. And the fact that the bug has been corrected has been detected by failure of routine checks during the weekly...



git-svn-id: svn://scm.gforge.inria.fr/svnroot/sollya/trunk@2279 af37bc7e-59aa-4313-bbaf-7d19c0032f32